requirements.txt
setuptools==68.0.0
matplotlib==3.7.1
numpy==1.24.2
pyfastnoisesimd==0.4.2
scipy==1.10.1
stable_baselines==2.0.0
sb3_contrib==2.0.0 # make sure to use whole sequence branch for speed up
torch==1.12.1+cu116
tensorboard==2.13.0
shimmy==1.1.0


-f https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu116